Moving randomexponential to test_diffusion and updating code to latest API
